% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{US cities}
\format{An \code{sf} layer of type \code{POINT}}
\usage{
cities
}
\description{
Point layer with US cities
}
\examples{
data(cities)
}
\keyword{datasets}
