% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_letter_to_num.R
\name{column_letter_to_num}
\alias{column_letter_to_num}
\title{Find the corresponding column number for an excel column letter}
\usage{
column_letter_to_num(x)
}
\arguments{
\item{x}{A vector of Excel column letters}
}
\value{
A vector of column numbers
}
\description{
Find the corresponding column number for an excel column letter
}
\examples{
column_letter_to_num("A")
column_letter_to_num("Z")
column_letter_to_num(c("Z", "AA"))
column_letter_to_num("BA")
column_letter_to_num("AAA")
column_letter_to_num(LETTERS)
}
\references{
\url{https://stackoverflow.com/questions/34537243/convert-excel-column-names-to-numbers}
}
