% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_num_to_letter.R
\name{column_num_to_letter}
\alias{column_num_to_letter}
\title{Find the corresponding column letter for a column number in Excel}
\usage{
column_num_to_letter(x)
}
\arguments{
\item{x}{A vector of Excel column numbers}
}
\value{
A vector of column letters
}
\description{
Find the corresponding column letter for a column number in Excel
}
\examples{
column_num_to_letter(1)
column_num_to_letter(26)
column_num_to_letter(c(26, 27))
column_num_to_letter(53)
column_num_to_letter(703)
column_num_to_letter(1:26)
}
\references{
\url{https://stackoverflow.com/questions/34537243/convert-excel-column-names-to-numbers}
}
