% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_download.R
\name{gtfs_download}
\alias{gtfs_download}
\title{Downloads and unzips GTFS files}
\usage{
gtfs_download(url, dir)
}
\arguments{
\item{url}{URL of the GTFS zip file}

\item{dir}{Path to destination directory}
}
\description{
Downloads and unzips GTFS files
}
\examples{
\dontrun{
url = "ftp://199.203.58.18/israel-public-transportation.zip"
dir = "/home/michael/Downloads/gtfs_test"
gtfs_download(url, dir)
list.files(dir)
}
}
