% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_read.R
\name{gtfs_read}
\alias{gtfs_read}
\title{Read GTFS files}
\usage{
gtfs_read(dir)
}
\arguments{
\item{dir}{Path to directory with \code{*.txt} GTFS files}
}
\value{
\code{list} where each element is a \code{data.frame} created from a GTFS \code{*.txt} file
}
\description{
Read GTFS files
}
\examples{
dir = "~/Downloads/gtfs"
dat = gtfs_read(dir)
names(dat)
lapply(dat, head)
}
