% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{itm7}
\alias{itm7}
\title{7-Parameter ITM projection}
\format{\code{character}}
\usage{
itm7
}
\description{
Proj string of the 7-Parameter ITM projection
}
\examples{
data(itm7)
}
\references{
\url{https://www.systematics.co.il/wp-content/uploads/ITM2WGS84_Transformation1.pdf}
}
\keyword{datasets}
