% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCircFilter.R
\name{makeCircFilter}
\alias{makeCircFilter}
\title{Function to make a circular weights matrix of given radius and resolution}
\usage{
makeCircFilter(radius, res)
}
\arguments{
\item{radius}{}

\item{res}{}
}
\value{

}
\description{
Function to make a circular weights matrix of given radius and resolution
}
\note{
NB radius must me an even multiple of res!
}
\examples{
m = make_circ_filter(400, 30) # filer for 400-m buffer on a 30-m resolution raster
image(m, asp = 1)
}
\references{
https://scrogster.wordpress.com/2012/10/05/applying-a-circular-moving-window-filter-to-raster-data-in-r/
}

