% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_combine.R
\name{pdf_combine}
\alias{pdf_combine}
\title{Combines separate PDF files into one page}
\usage{
pdf_combine(
  input,
  nrow,
  ncol,
  height = 11.7,
  width = 8.3,
  output = "combined.pdf"
)
}
\arguments{
\item{input}{A vector of PDF file names or paths}

\item{nrow}{Number of rows in output}

\item{ncol}{Number of columns in output}

\item{height}{Output file height, in inches (default is \code{11.7})}

\item{width}{Output file width, in inches (default is \code{8.3})}

\item{output}{The name of the combined PDF file name}
}
\description{
Combines PDF files to into one file/page. Requires \code{pdfjam} command-line utility.
}
\examples{
\dontrun{
}
}
