% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_expand.R
\name{pdf_expand}
\alias{pdf_expand}
\title{Expand PDF page size}
\usage{
pdf_expand(filename)
}
\arguments{
\item{filename}{A PDF file name or file path}
}
\description{
Expand the page size of a PDF file to include all of its contents. Requires \code{inkscape} command-line utility.
}
\details{
The function modifies an existing PDF file!
}
\examples{
\dontrun{
pdf_expand(filename = "~/test.pdf")
}
}
