% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_split.R
\name{pdf_split}
\alias{pdf_split}
\title{Split PDF to separate files}
\usage{
pdf_split(filename)
}
\arguments{
\item{filename}{A PDF file name or file path}
}
\description{
Split a PDF file to separate files, one file per page. Requires \code{pdftk} command-line utility.
}
\examples{
\dontrun{
pdf_split(filename = "~/test.pdf")
}
}
