% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_png.R
\name{pdf_to_png}
\alias{pdf_to_png}
\title{Convert PDF to PNG}
\usage{
pdf_to_png(filename, dpi = 300)
}
\arguments{
\item{filename}{A PDF file name or file path}

\item{dpi}{Required DPI}
}
\description{
Convert a PDF file to a PNG file. Requires \code{inkscape} command-line utility.
}
\details{
The function creates an PNG file at the same location where the PDF is.
}
\examples{
\dontrun{
pdf_to_png(filename = "~/test.pdf")
}
}
