% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polToSeg.R
\name{polToSeg}
\alias{polToSeg}
\title{Polygons or lines to segments}
\usage{
polToSeg(x)
}
\arguments{
\item{x}{A \code{SpatialLines*} or a \code{SpatialPolygons*} object.}
}
\value{
A \code{SpatialLines} object where each segment is represented by a separate feature.
}
\description{
Split lines or polygons to separate segments.
}
\examples{
# Create sample SpatialPoygons named 'SpP'
Sr1 = Polygon(cbind(c(2,4,4,1,2),c(2,3,5,4,2)))
Sr2 = Polygon(cbind(c(5,4,2,5),c(2,3,2,2)))
Sr3 = Polygon(cbind(c(4,4,5,10,4),c(5,3,2,5,5)))
Sr4 = Polygon(cbind(c(5,6,6,5,5),c(4,4,3,3,4)), hole = TRUE)
Srs1 = Polygons(list(Sr1), "s1")
Srs2 = Polygons(list(Sr2), "s2")
Srs3 = Polygons(list(Sr3, Sr4), "s3/4")
SpP = SpatialPolygons(list(Srs1,Srs2,Srs3), 1:3)
# Convert 'SpP' to segments
seg = polToSeg(SpP)
plot(seg, col = sample(rainbow(15)))
text(gCentroid(seg, byid = TRUE), add = TRUE)

}
\references{
The function is based on code from the following 'r-sig-geo' post by Roger Bivand:

\url{https://stat.ethz.ch/pipermail/r-sig-geo/2013-April/017998.html}
}

