% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{Simple word capitalization}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A capitalized character vector
}
\description{
Simple word capitalization
}
\examples{
name = c("zip code", "state", "final count")
simpleCap(name)
}
\references{
\url{https://stackoverflow.com/questions/6364783/capitalize-the-first-letter-of-both-words-in-a-two-word-string}
}
