% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{sp.get.shortest.path}
\alias{sp.get.shortest.path}
\title{Title}
\usage{
sp.get.shortest.path(sln, from, to, fromColumn = "start", toColumn = "end")
}
\arguments{
\item{sl}{}
}
\value{
sln
}
\description{
Title
}
\examples{

}
