% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{US states}
\format{
An \code{sf} layer of type \code{POLYGON}
}
\usage{
states
}
\description{
Polygonal layer with US states
}
\examples{
data(states)

}
\keyword{datasets}
