% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{voronoi}
\alias{voronoi}
\title{Voronoi polygons within given extent}
\usage{
voronoi(pnt, pol)
}
\arguments{
\item{pnt}{points (locations) of the queries}

\item{pol}{layer from which the geometries or attributes are queried}
}
\value{
Voronoi polygons (\code{SpatialPolygons}).
}
\description{
Create voronoi polygons, given points and extent
}
\examples{
library(rgdal)
data(states)
data(cities)
usAtlas = "+proj=laea +lat_0=45 +lon_0=-100 +x_0=0 +y_0=0 +a=6370997 +b=6370997 +units=m +no_defs"
states = spTransform(states, usAtlas)
cities = spTransform(cities, usAtlas)
cities = cities[states, ]
plot(states)
plot(cities, add = TRUE)
v = voronoi(cities, states)
plot(v, add = TRUE, border = "red")
}
\references{
http://stackoverflow.com/questions/12156475/combine-voronoi-polygons-and-maps
}

