% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{zone}
\alias{zone}
\title{Determine UTM zone}
\usage{
zone(x)
}
\arguments{
\item{x}{An \code{sf}, \code{sgc} or \code{sfg} object}
}
\value{
EPSG code (\code{numeric}) which can be passed to other functions such as \code{st_transform}
}
\description{
The function accepts a vector layer (\code{sf}, \code{sgc} or \code{sfg}) and returns the EPSG code of the UTM zone according to the centroid of the bounding box.
}
\examples{
zone(cities[cities$name == "Boston MA", ])
}
